// src/app/models/employee.model.ts
export class Employee {
  id?: number;
  name?: string;
  gender?: string;
  email?: string;
  fullname?: string;
  phoneNumber?: number;
  contactPreference?: string;
  dateOfBirth?: Date;
  department?: string;
  isActive?: boolean;
  photoPath?: string;
  confirmPassword?: string;
  password?: string;
  
}